Allowed home environment variable to be set
This was fixed in the package provider, but never fixed in the exec the nodejs::npm class was utilizing for installations:

Heres a bit of background:

Node-gyp expects either HOME or USERPROFILE to be defined in the environment. If it is not set, installation will fail.

```
Error: node-gyp requires that the user's home directory is specified in either of the environmental variables HOME or USERPROFILE
```

Update nodejs_npm_spec.rb

Added spec test where the environment variable HOME is defined as '/root'

Update nodejs_npm_spec.rb

fixed spec test

fixed whitespacing issues